[GNA] Enable int8 weights for convolutions
