+++ src/plugins/intel_gna/src/backend/pwl_border_values_counter_identity.cpp
@@ -7,7 +7,7 @@
 
 #include "gna_plugin_log.hpp"
 #include "pwl_input_params.hpp"
-#include "round_float_define.hpp"
+#include "common/numerical_utils.hpp"
 
 namespace ov {
 namespace intel_gna {
