+++ src/plugins/intel_gna/src/frontend/model_quantizer.hpp
@@ -10,102 +10,67 @@
 #include <type_traits>
 
 #include <legacy/details/ie_cnn_network_tools.h>
-#include <legacy/layer_transform.hpp>
 
 #include "gna_graph_tools.hpp"
+#include "gna_plugin_config.hpp"
 #include "layer_quantizer.hpp"
 #include "scale_factor_calc.hpp"
 #include "weights_converter.hpp"
 #include "gna_itt.hpp"
 #include "descriptions/gna_desc.hpp"
 
-namespace GNAPluginNS {
+namespace ov {
+namespace intel_gna {
+namespace frontend {
 
 /**
  * Quantize entire network
  */
 class ModelQuantizer {
- public:
-    // Below four functions are used only in unit tests
-    InferenceEngine::CNNNetwork quantize(const InferenceEngine::CNNNetwork& model, float scale_factor) const {
-        return quantize(model, [](const InferenceEngine::CNNNetwork&, bool run_before_copy, bool low_precision){}, scale_factor);
-    }
+    const Config& gna_config;
+    const bool fake_quantized;
 
-    template <class PreQuantisationCb>
-    InferenceEngine::CNNNetwork quantize(const InferenceEngine::CNNNetwork& model, const PreQuantisationCb& cb, float scale_factor) const {
-        return quantize(model, cb, std::vector<float>({scale_factor}));
-    }
-
-    InferenceEngine::CNNNetwork quantize(const InferenceEngine::CNNNetwork& model, std::vector<float> scale_factors) const {
-        return quantize(model, [](InferenceEngine::CNNNetwork&, bool run_before_copy, bool low_precision){}, scale_factors);
-    }
-
-    template <class PreQuantisationCb>
-    InferenceEngine::CNNNetwork quantize(const InferenceEngine::CNNNetwork& model,  const PreQuantisationCb& cb, std::vector<float> scale_factors) const {
-        GNAPluginNS::GnaInputs inputs;
-        InferenceEngine::InputsDataMap inputs_map = model.getInputsInfo();
-        int sf_id = 0;
-        for (auto &&input_data : inputs_map) {
-            auto input_layer = getCreatorLayer(input_data.second->getInputData()).lock();
-            if (scale_factors.size() <= sf_id) {
-                THROW_GNA_EXCEPTION << "Scale factors are not set for some of the inputs";
-            }
-            inputs[input_layer->name].scale_factor = scale_factors[sf_id++];
-        }
-
-        return quantize(model, cb, inputs, InferenceEngine::Precision::I16, false);
-    }
-
-    // This function is acutally used by the plugin
+public:
+    ModelQuantizer(const Config& gna_config, const bool fake_quantized)
+        : gna_config(gna_config),
+          fake_quantized(fake_quantized) {}
     template <class PreQuantisationCb>
     InferenceEngine::CNNNetwork quantize(const InferenceEngine::CNNNetwork& model,
                                          const PreQuantisationCb& cb,
-                                         const GNAPluginNS::GnaInputs& inputs,
-                                         const InferenceEngine::Precision& gna_precision,
-                                         const bool& inputs_int8_precision) const {
+                                         const GnaInputs& inputs) const {
         OV_ITT_SCOPED_TASK(itt::domains::GNA_LT, "ModelQuantizer::quantize");
-        auto visitor = [&](InferenceEngine::CNNLayerPtr lp) {
-            auto new_layer = InferenceEngine::injectData<QuantizedLayerParams>(lp);
-            transformLayer(new_layer, WeightsConverter());
+        auto visitor = [&](InferenceEngine::CNNLayerPtr layer_ptr) {
+            auto new_layer = InferenceEngine::injectData<frontend::QuantizedLayerParams>(layer_ptr);
+            convertBlobs(*new_layer);
             return new_layer;
         };
 
-        // User's hints for GNA weights precision
-        const bool weights_int8_precision = (gna_precision == InferenceEngine::Precision::I8) ? true : false;
-
         InferenceEngine::CNNNetwork copied_net = InferenceEngine::CNNNetCopy(model);
-        cb(copied_net, true, inputs_int8_precision);
-
+        cb(copied_net, true, gna_config.gnaFlags.input_low_precision);
         copied_net = InferenceEngine::CNNNetCopy(copied_net, visitor);
 
         // Allow client code to access copied topology, to avoid copies if user would like to chain quantisation with
         // another preprocessing
-        cb(copied_net, false, inputs_int8_precision);
+        cb(copied_net, false, gna_config.gnaFlags.input_low_precision);
 
-        LayersQuantizer lc(GNAPluginNS::kScaleFactorDefault);
         auto sorted_new_net = InferenceEngine::details::CNNNetSortTopologically(copied_net);
         gnalog() << "Sorted layers: " << std::endl;
-        for (auto &&layer : sorted_new_net) {
-            auto quant_data = InferenceEngine::getInjectedData<QuantizedLayerParams>(layer);
-            quant_data->_inputs_int8_precision = inputs_int8_precision;
-            quant_data->_weights_int8_precision = weights_int8_precision;
-            gnalog() << layer->name << std::endl;
-        }
 
         // Filling scale factors for input layers, memory layers will have scale factor of 1.0 by default
         InferenceEngine::InputsDataMap dm = copied_net.getInputsInfo();
         for (auto &&inputData : dm) {
             auto input_layer = getCreatorLayer(inputData.second->getInputData()).lock();
-            auto quant_data = InferenceEngine::getInjectedData<QuantizedLayerParams>(input_layer);
+            auto quant_data = InferenceEngine::getInjectedData<frontend::QuantizedLayerParams>(input_layer);
             IE_ASSERT(quant_data != nullptr);
             quant_data->_src_quant.SetScale(inputs.at(input_layer->name).scale_factor);
         }
 
+        // Propagate scale factor and quantize layers
         propagateScaleFactor(sorted_new_net);
+        frontend::LayerQuantizer lq(gna_config);
 
-        // Sorted order gives possibility for propagate quantisation along depended layers
         for (auto &&layer : sorted_new_net) {
-            transformLayer(layer, lc);
+            lq.quantize(*layer);
         }
 
         return copied_net;
@@ -113,8 +78,7 @@ class ModelQuantizer {
 
  private :
     void propagateScaleFactor(std::vector<InferenceEngine::CNNLayerPtr>& net) const {
-        ScaleFactorCalculator sf(net);
-
+        ScaleFactorCalculator sf(net, gna_config, fake_quantized);
         uint32_t inf_loop_count = 0;
         std::vector<std::string> inf_loop_pattern;
         std::vector<std::string> inf_loop_history;
@@ -122,7 +86,7 @@ class ModelQuantizer {
             auto layers = sf.getStartLayers();
             inf_loop_history.emplace_back(layers.front()->name);
             for (auto &&layer : layers) {
-                transformLayer(layer, sf);
+                sf.CalculateScaleFactor(layer);
                 // transforming until we reached cases where output scale updated due to situation in downstream layer
                 if (sf.needToRestart()) {
                     inf_loop_history.back() += "#" + layer->name;
@@ -207,4 +171,7 @@ class ModelQuantizer {
         }
     }
 };
-}  // namespace GNAPluginNS
+
+}  // namespace frontend
+}  // namespace intel_gna
+}  // namespace ov
