+++ src/plugins/intel_gna/src/frontend/quantization.cpp
@@ -5,10 +5,25 @@
 #include <cstring>
 #include <gna_plugin_log.hpp>
 #include <limits>
+#include <algorithm>
+#include "layers/gna_fake_quantize_layer.hpp"
 #include "backend/gna_types.h"
 #include "quantization.hpp"
-#include <algorithm>
 
+namespace ov {
+namespace intel_gna {
+namespace frontend {
+
+float ApplyFQ(float value, float input_low, float input_high, float output_low, float output_high, uint32_t levels) {
+    if (value <= std::min(input_low, input_high)) {
+        return output_low;
+    } else if (value > std::max(input_low, input_high)) {
+        return output_high;
+    } else {
+        return nearbyint((value - input_low) / (input_high - input_low) * (levels - 1)) /
+            (levels - 1) * (output_high - output_low) + output_low;
+    }
+}
 
 std::pair<float, float> FindMinMaxValues(void* ptr_float_memory, size_t num_elements) {
     float* ptr_float_feat = reinterpret_cast<float*>(ptr_float_memory);
@@ -118,15 +133,7 @@ void QuantizeWeights<int8_t>(const QuantizationData& data, float* ptr_float_weig
             float value = ptr_float_weights[offset];
 
             if (min_values_size > 0) {
-                auto x = value;
-                if (x <= std::min(input_low, input_high)) {
-                    value = output_low;
-                } else if (x > std::max(input_low, input_high)) {
-                    value = output_high;
-                } else {
-                    value = nearbyint((x - input_low) / (input_high - input_low) * (levels - 1)) / (levels - 1) *
-                                (output_high - output_low) + output_low;
-                }
+                value = ApplyFQ(value, input_low, input_high, output_low, output_high, levels);
             }
 
             if (ptr_int_biases) {
@@ -141,15 +148,7 @@ void QuantizeWeights<int8_t>(const QuantizationData& data, float* ptr_float_weig
 
             int8_t* ptr_weight_8 = ptr_int_weights + offset;
 
-            if (value > std::numeric_limits<int8_t>::max()) {
-                *ptr_weight_8 = std::numeric_limits<int8_t>::max();
-                num_saturate++;
-            } else if (value < std::numeric_limits<int8_t>::min()) {
-                *ptr_weight_8 = std::numeric_limits<int8_t>::min();
-                num_saturate++;
-            } else {
-                *ptr_weight_8 = static_cast<int8_t>(value);
-            }
+            *ptr_weight_8 = SaturationCast<int8_t>(value, &num_saturate);
         }
     }
 
@@ -187,30 +186,14 @@ void QuantizeWeights<int16_t>(const QuantizationData& data, float* ptr_float_wei
             float rounding_value = (ptr_float_weights[row * data.num_columns + col] > 0) ? 0.5f : -0.5f;
             float value = ptr_float_weights[row * data.num_columns + col];
             if (min_values_size > 0) {
-                auto x = value;
-                if (x <= std::min(input_low, input_high)) {
-                    value = output_low;
-                } else if (x > std::max(input_low, input_high)) {
-                    value = output_high;
-                } else {
-                    value = nearbyint((x - input_low) / (input_high - input_low) * (levels - 1)) /
-                        (levels - 1) * (output_high - output_low) + output_low;
-                }
+                value = ApplyFQ(value, input_low, input_high, output_low, output_high, levels);
             }
 
             value = value * data.scale_factor + rounding_value;
 
             int16_t* ptr_weight_16 = ptr_int_weights + (row * data.num_columns + col);
 
-            if (value > std::numeric_limits<int16_t>::max()) {
-                *ptr_weight_16 = std::numeric_limits<int16_t>::max();
-                num_saturate++;
-            } else if (value < std::numeric_limits<int16_t>::min()) {
-                *ptr_weight_16 = std::numeric_limits<int16_t>::min();
-                num_saturate++;
-            } else {
-                *ptr_weight_16 = (int16_t)value;
-            }
+            *ptr_weight_16 = SaturationCast<int16_t>(value, &num_saturate);
         }
     }
 
@@ -239,15 +222,8 @@ void QuantizeBiases<int32_t>(const QuantizationData& data, float* ptr_float_bias
         for (size_t row = 0; row < data.num_rows; row++) {
             float rounding_value = (ptr_float_biases[row] > 0) ? 0.5f : -0.5f;
             float value = ptr_float_biases[row] * data.scale_factor + rounding_value;
-            if (value > std::numeric_limits<int32_t>::max()) {
-                ptr_int_biases[row] = std::numeric_limits<int32_t>::max();
-                num_saturate++;
-            } else if (value < std::numeric_limits<int32_t>::min()) {
-                ptr_int_biases[row] = std::numeric_limits<int32_t>::min();
-                num_saturate++;
-            } else {
-                ptr_int_biases[row] = static_cast<int32_t>(value);
-            }
+
+            ptr_int_biases[row] = SaturationCast<int32_t>(value, &num_saturate);
         }
     }
 
@@ -265,15 +241,8 @@ void QuantizeBiases<gna_compound_bias_t>(const QuantizationData& data, float* pt
         for (size_t row = 0; row < data.num_rows; row++) {
             float rounding_value = (ptr_float_biases[row] > 0) ? 0.5f : -0.5f;
             float value = ptr_float_biases[row] * data.scale_factor + rounding_value;
-            if (value > std::numeric_limits<int32_t>::max()) {
-                ptr_int_biases[row].bias = std::numeric_limits<int32_t>::max();
-                num_saturate++;
-            } else if (value < std::numeric_limits<int32_t>::min()) {
-                ptr_int_biases[row].bias = std::numeric_limits<int32_t>::min();
-                num_saturate++;
-            } else {
-                ptr_int_biases[row].bias = static_cast<int32_t>(value);
-            }
+
+            ptr_int_biases[row].bias = SaturationCast<int32_t>(value, &num_saturate);
         }
     }
     if (num_saturate > 0) {
@@ -281,3 +250,7 @@ void QuantizeBiases<gna_compound_bias_t>(const QuantizationData& data, float* pt
                  << " saturations in compound biases quantization." << std::endl;
     }
 }
+
+}  // namespace frontend
+}  // namespace intel_gna
+}  // namespace ov
