+++ src/plugins/intel_gna/src/layers/gna_fake_quantize_layer.hpp
@@ -11,6 +11,8 @@
 
 #include <ie_algorithm.hpp>
 
+using namespace ov::intel_gna;
+
 namespace GNAPluginNS {
 class GNAFakeQuantizeLayer {
     InferenceEngine::CNNLayerPtr fqLayer;
@@ -23,7 +25,7 @@ class GNAFakeQuantizeLayer {
     }
 
     /**
-     * @brief convert FQ layer directly to gna-pwl activation layer
+     * @brief Convert FQ layer directly to gna-pwl activation layer
      */
     DnnActivation parseAsActivation() const {
         DnnActivation fqActivation{};
@@ -49,16 +51,17 @@ class GNAFakeQuantizeLayer {
 
         return fqActivation;
      }
-     /**
-      * retrieves input blob for FQ layer that connected to const layer
-      */
-     InferenceEngine::Blob::Ptr getConstInputData() const {
-         return LayerUtils::getParamFromInputAsBlob(fqLayer, 0);
-     }
 
-     /**
-      * fake quantize has 5 input layers, while 4 of them always constant layer, and 1 might be a tensor - connection
-      */
+    /**
+     * @brief Retrieve input blob for FQ layer that connected to const layer
+     */
+    InferenceEngine::Blob::Ptr getConstInputData() const {
+        return LayerUtils::getParamFromInputAsBlob(fqLayer, 0);
+    }
+
+    /**
+     * @brief Fake quantize has 5 input layers, while 4 of them always constant layer, and 1 might be a tensor - connection
+     */
     InferenceEngine::CNNLayerPtr getInputLayer() const {
         return getInputLayerAt(fqLayer, 0);
     }
@@ -101,10 +104,10 @@ class GNAFakeQuantizeLayer {
             break;
         }
         case InferenceEngine::Precision::FP16: {
-            auto dataMinFP32 = make_fp32_blob(dataMin);
+            auto dataMinFP32 = frontend::make_fp32_blob(dataMin);
             memcpy(&minValues[0], dataMinFP32->buffer().as<float*>(), rangeSize * sizeof(float));
 
-            auto dataMaxFP32 = make_fp32_blob(dataMax);
+            auto dataMaxFP32 = frontend::make_fp32_blob(dataMax);
             memcpy(&maxValues[0], dataMaxFP32->buffer().as<float*>(), rangeSize * sizeof(float));
             break;
         }
