+++ src/plugins/intel_gna/src/memory/gna_memory_state.cpp
@@ -4,9 +4,9 @@
 
 #include "gna_memory_state.hpp"
 #include "frontend/quantized_layer_params.hpp"
-#include <legacy/layer_transform.hpp>
 #include "preprocessing.hpp"
 #include "ie_layouts.h"
+#include "gna_graph_tools.hpp"
 
 namespace  GNAPluginNS {
 
@@ -66,7 +66,8 @@ namespace memory {
         switch (state_precision) {
         case InferenceEngine::Precision::I16: {
             if (new_state_precision == InferenceEngine::Precision::FP32) {
-                auto quantized = InferenceEngine::getInjectedData<QuantizedLayerParams>(state->getInput());
+                auto quantized =
+                    InferenceEngine::getInjectedData<ov::intel_gna::frontend::QuantizedLayerParams>(state->getInput());
                 auto scale_factor = quantized != nullptr ? quantized->_dst_quant.GetScale() : state->scale_factor;
                 GNAPluginNS::ConvertToInt16(static_cast<int16_t*>(state->gna_ptr),
                     newState->buffer().as<float*>(),
@@ -92,7 +93,8 @@ namespace memory {
         InferenceEngine::Precision state_precision = getPrecision();
 
         if (state->getInput() && state_precision == InferenceEngine::Precision::I16) {
-            auto quantized = InferenceEngine::getInjectedData<QuantizedLayerParams>(state->getInput());
+            auto quantized =
+                InferenceEngine::getInjectedData<ov::intel_gna::frontend::QuantizedLayerParams>(state->getInput());
             auto scale_factor = quantized != nullptr ? quantized->_dst_quant.GetScale() : state->scale_factor;
 
             auto result_blob = make_blob_with_precision(InferenceEngine::TensorDesc(InferenceEngine::Precision::FP32,
@@ -119,7 +121,8 @@ namespace memory {
     }
 
     float GNAVariableState::GetScaleFactor() const {
-        auto quantized = InferenceEngine::getInjectedData<QuantizedLayerParams>(state->getInput());
+        auto quantized =
+            InferenceEngine::getInjectedData<ov::intel_gna::frontend::QuantizedLayerParams>(state->getInput());
         auto scale_factor = quantized != nullptr ? quantized->_dst_quant.GetScale() : state->scale_factor;
         return scale_factor;
     }
