+++ src/plugins/intel_gna/src/runtime/pwl.cpp
@@ -9,6 +9,7 @@
 #include <limits>
 #include <cstdint>
 #include <algorithm>
+#include "frontend/quantization.hpp"
 
 #ifdef _NO_MKL_
 #include <cmath>
@@ -27,7 +28,7 @@
 #include "pwl.h"
 #include "gna_plugin_log.hpp"
 #include "gna_slope_scale.h"
-#include "round_float_define.hpp"
+#include "common/numerical_utils.hpp"
 #include "ops/reference/pwl.hpp"
 
 double relu(const double x) { if (x < 0) { return(0.0); } else { return(x); } }
@@ -330,17 +331,13 @@ void PwlDesign(const DnnActivation& activation_type,
                 gnalog() << "=========================== Pow Segments===========================\n";
                 uint32_t num_segment_size = 0;
 
-                auto fp32eq = [](float p1, float p2) -> bool {
-                    return (std::abs(p1 - p2) <= 0.00001f * std::min(std::abs(p1), std::abs(p2)));
-                };
-
                 auto args = std::tuple<double, double, double>{ activation_type.args.pow.exponent,
                                                                 activation_type.args.pow.scale,
                                                                 activation_type.args.pow.offset };
 
                 auto input_min_value = static_cast<double>(std::numeric_limits<int32_t>::min());
                 auto input_max_value = static_cast<double>(std::numeric_limits<int32_t>::max());
-                double x_min = fp32eq(fmod(activation_type.args.pow.exponent, 1.0), 0.0f)? input_min_value / scale_in: 0.0;
+                double x_min = ov::intel_gna::common::fp32eq(fmod(activation_type.args.pow.exponent, 1.0), 0.0f)? input_min_value / scale_in: 0.0;
                 x_min = std::max(x_min, -POW_DOMAIN);
 
                 double x_max = input_max_value / scale_in;
@@ -530,16 +527,8 @@ void PwlApply32(intel_dnn_component_t *component,
 
                 for (uint32_t j = num_col_start; j <= num_col_end; j++) {
                     auto offset = i * num_columns + j;
-                    auto x = ptr_in[offset];
 
-                    if (x <= std::min(input_low, input_high)) {
-                        ptr_out[offset] = static_cast<float>(output_low);
-                    } else if (x > std::max(input_low, input_high)) {
-                        ptr_out[offset] = static_cast<float>(output_high);
-                    } else {
-                        ptr_out[offset] = static_cast<float>(nearbyint((x - input_low) / (input_high - input_low) * (levels - 1)) /
-                            (levels - 1) * (output_high - output_low) + output_low);
-                    }
+                    ptr_out[offset] = ov::intel_gna::frontend::ApplyFQ(ptr_in[offset], input_low, input_high, output_low, output_high, levels);
                 }
             }
             break;
