+++ src/plugins/intel_gna/tests/unit/gna_hw_precision_test.cpp
@@ -26,6 +26,7 @@ public:
     }
     void set_low_precision_input() {
         this->gnaFlags->input_low_precision = true;
+        this->config.gnaFlags.input_low_precision = true;
     }
 };
