+++ src/tests_deprecated/unit/engines/gna/matchers/diag_matcher.hpp
@@ -4,7 +4,6 @@
 
 #pragma once
 
-#include "backend/gna_types.h"
 #include "nnet_base_matcher.hpp"
 #include "frontend/quantization.hpp"
