+++ src/tests_deprecated/unit/engines/gna/matchers/weights_matcher.hpp
@@ -3,9 +3,7 @@
 //
 
 #pragma once
-#include "backend/gna_types.h"
 #include "nnet_base_matcher.hpp"
-#include "frontend/quantization.hpp"
 
 using TranspozedData = std::tuple<std::vector<uint16_t>*, int, int>;
